sql_schema_json_validator/requirements.txt
pyyaml
SQLAlchemy>=2.0
psycopg2-binary>=2.9
